requirements.txt
httpx>=0.27.0
python-dotenv>=1.0.1
uvloop>=0.20.0 ; sys_platform != "win32"
zhipuai>=2.0.0
